+++ uui-components/src/overlays/DropdownContainer.tsx
@@ -3,11 +3,11 @@ import FocusLock from 'react-focus-lock';
 import {
     uuiElement, IHasCX, IHasChildren, cx, IHasRawProps, uuiMarkers, IHasForwardedRef, IDropdownBodyProps,
     IHasStyleAttrs,
-    useUuiContext,
 } from '@epam/uui-core';
 import { VPanel } from '../layout/flexItems/VPanel';
 import PopoverArrow from './PopoverArrow';
 import { ReactFocusLockProps } from 'react-focus-lock';
+import { useCallback, useState } from 'react';
 
 export interface DropdownContainerProps
     extends IHasCX,
@@ -57,13 +57,18 @@ export const DropdownContainer = React.forwardRef((props: DropdownContainerProps
         returnFocus = true,
         persistentFocus = true,
     } = props;
+    const [shadowRootHost, setShadowRootHost] = useState<ShadowRoot | null>(null);
 
-    const { shadowRootHost } = useUuiContext();
+    const saveShadowRootHost = useCallback((node: HTMLElement | null) => {
+        if (!node) return;
+        const root = node.getRootNode();
+        setShadowRootHost(root instanceof ShadowRoot ? root : null);
+    }, []);
 
     function renderDropdownContainer() {
         return (
             <VPanel
-                forwardedRef={ !focusLock ? ref as React.ForwardedRef<HTMLDivElement> : undefined }
+                forwardedRef={ !focusLock ? (ref as React.ForwardedRef<HTMLDivElement>) : saveShadowRootHost }
                 cx={ cx(uuiElement.dropdownBody, uuiMarkers.lockFocus, props.cx) }
                 style={ {
                     ...props.style,